server/agents/feedback/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
aiosqlite
requests
mistralai